+++ built/grammar/AbstractRuntimeNonRootNode.js
@@ -17,6 +17,7 @@ class AbstractRuntimeNonRootNode extends AbstractRuntimeNode_1.default {
     }
     getDefinition() {
         // todo: do we need a relative to with this firstWord path?
+        console.log(this.getFirstWordPath());
         return this._getNodeTypeDefinitionByName(this.getFirstWordPath());
     }
     getConstantsObject() {