+++ releaseNotes.md
@@ -18,6 +18,7 @@ edge / 2019-06-11
 - Breaking: getRunTimeNodeTypeNames is now getRunTimeNodeTypeIds
 - Breaking: getFirstCellType is now getFirstCellTypeId
 - Breaking: getCellTypeDefinition is now getCellTypeDefinitionById
+- Breaking: getNodeTypeDefinitionByName is now getNodeTypeDefinitionByNodeTypeId
 - Infra: test coverage
 
 25.2.0 / 2019-05-30